<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Adding Symbol.Canonical property which will be cached. Adding unit and
  regression tests

Master
```
73.52 seconds at 739k data points per second. Processing total of 54,299,839 data points.
74.59 seconds at 728k data points per second. Processing total of 54,299,839 data points.
72.02 seconds at 754k data points per second. Processing total of 54,299,839 data points.
```
PR
```
65.05 seconds at 835k data points per second. Processing total of 54,299,839 data points.
64.15 seconds at 846k data points per second. Processing total of 54,299,839 data points.
65.60 seconds at 828k data points per second. Processing total of 54,299,839 data points.
```
Performance algorithm used
```
class BasicTemplateOptionsAlgorithm(QCAlgorithm):
    def Initialize(self):
        self.SetStartDate(2010, 1, 1)
        self.SetEndDate(2011, 1, 1)

        equity = self.AddEquity("AAPL")
        option = self.AddOption("AAPL")
        self.option_symbol = option.Symbol

        option.SetFilter(lambda u: (u.Strikes(-2, +2)
                                     .Expiration(0, 180)))
        self.SetBenchmark(equity.Symbol)
```
#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/4732
Related to https://github.com/QuantConnect/Lean/issues/5343
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Improve performance and user experience
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit and regression tests. Performance tests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [x] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->